tools: don't attempt to overwrite kconfig if checking fails

+++ format.sh
@@ -30,6 +30,7 @@ fi
 KCONFIG_FILE=main/Kconfig.projbuild
 
 if [ -x ${IDF_PATH}/tools/ci/check_kconfigs.py ]; then
+    rm -f ${KCONFIG_FILE}.new
     ${IDF_PATH}/tools/ci/check_kconfigs.py ${KCONFIG_FILE} || true
-    mv ${KCONFIG_FILE}.new ${KCONFIG_FILE}
+    [ -f ${KCONFIG_FILE}.new ] && mv ${KCONFIG_FILE}.new ${KCONFIG_FILE}
 fi
